browser/extensions/formautofill/test/mochitest/test_basic_creditcard_autocomplete_form.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test basic autofill</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="formautofill_common.js"></script>
  <script type="text/javascript" src="satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Form autofill test: simple form credit card autofill

<script>
/* import-globals-from ../../../../../testing/mochitest/tests/SimpleTest/SpawnTask.js */
/* import-globals-from ../../../../../toolkit/components/satchel/test/satchel_common.js */
/* import-globals-from formautofill_common.js */

"use strict";

const {FormAutofillUtils} = SpecialPowers.Cu.import("resource://formautofill/FormAutofillUtils.jsm");

let MOCK_STORAGE = [{
  "cc-name": "John Doe",
  "cc-number": "1234567812345678",
  "cc-exp-month": 4,
  "cc-exp-year": 2017,
}, {
  "cc-name": "Timothy Berners-Lee",
  "cc-number": "1111222233334444",
  "cc-exp-month": 12,
  "cc-exp-year": 2022,
}];

function checkElementFilled(element, expectedvalue) {
  return [
    new Promise(resolve => {
      element.addEventListener("input", function onInput() {
        ok(true, "Checking " + element.name + " field fires input event");
        resolve();
      }, {once: true});
    }),
    new Promise(resolve => {
      element.addEventListener("change", function onChange() {
        ok(true, "Checking " + element.name + " field fires change event");
        is(element.value, expectedvalue, "Checking " + element.name + " field");
        resolve();
      }, {once: true});
    }),
  ];
}

function checkAutoCompleteInputFilled(element, expectedvalue) {
  return new Promise(resolve => {
    element.addEventListener("DOMAutoComplete", function onChange() {
      is(element.value, expectedvalue, "Checking " + element.name + " field");
      resolve();
    }, {once: true});
  });
}

function checkFormFilled(address) {
  let promises = [];
  for (let prop in address) {
    let element = document.getElementById(prop);
    if (document.activeElement == element) {
      promises.push(checkAutoCompleteInputFilled(element, address[prop]));
    } else {
      let converted = address[prop];
      if (prop == "street-address") {
        converted = FormAutofillUtils.toOneLineAddress(converted);
      }
      promises.push(...checkElementFilled(element, converted));
    }
  }
  doKey("return");
  return Promise.all(promises);
}

async function setupAddressStorage() {
  await addCreditCard(MOCK_STORAGE[0]);
  await addCreditCard(MOCK_STORAGE[1]);
}

// TODO: Add form history fallback test cases for credit card fields. (bug 1393374)
async function setupFormHistory() {
  await updateFormHistory([
    {op: "add", fieldname: "cc-name", value: "John Smith"},
    {op: "add", fieldname: "cc-exp-year", value: 2023},
  ]);
}

initPopupListener();

// Disable insecure warning for credit card.
add_task(async function history_only_menu_checking() {
  await SpecialPowers.pushPrefEnv({
    set: [
      ["security.insecure_field_warning.contextual.enabled", false],
      ["extensions.formautofill.loglevel", "Debug"],
    ],
  });
});

// Form with history only.
add_task(async function history_only_menu_checking() {
  await setupFormHistory();

  await setInput("#cc-exp-year", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(["2023"], false);
});

// Display history search result if less than 3 inputs are covered by all saved
// fields in the storage.
add_task(async function all_saved_fields_less_than_threshold() {
  const mockCreditCard = {
    "cc-name": "test",
    "cc-number": "1234123456785678",
  };

  await addCreditCard(mockCreditCard);

  await setInput("#cc-name", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries([JSON.stringify({
    primary: "test",
    secondary: "****5678",
  })], true);

  await cleanUpCreditCards();
});

/*
// Form with both history and address storage.
add_task(async function check_menu_when_both_existed() {
  await setupAddressStorage();

  await setInput("#organization", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: address.organization,
      secondary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
    })
  ));

  await setInput("#street-address", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
      secondary: address.organization,
    })
  ));

  await setInput("#tel", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: address.tel,
      secondary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
    })
  ));

  await setInput("#address-line1", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
      secondary: address.organization,
    })
  ));
});

// Display history search result if no matched data in addresses.
add_task(async function check_fallback_for_mismatched_field() {
  await setInput("#email", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(["foo@mozilla.com"], false);
});

// Display history search result if address autofill is disabled.
add_task(async function check_search_result_for_pref_off() {
  await SpecialPowers.pushPrefEnv({
    set: [["extensions.formautofill.addresses.enabled", false]],
  });

  await setInput("#tel", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(["+1234567890"], false);

  await SpecialPowers.popPrefEnv();
});

// Autofill the address from dropdown menu.
add_task(async function check_fields_after_form_autofill() {
  await setInput("#organization", "Moz");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: address.organization,
      secondary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
    })
  ).slice(1));
  doKey("down");
  await checkFormFilled(MOCK_STORAGE[1]);
});

// Fallback to history search after autofill address.
add_task(async function check_fallback_after_form_autofill() {
  await setInput("#tel", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(["+1234567890"], false);
});

// Resume form autofill once all the autofilled fileds are changed.
add_task(async function check_form_autofill_resume() {
  document.querySelector("#tel").blur();
  document.querySelector("#form1").reset();
  await setInput("#tel", "");
  doKey("down");
  await expectPopup();
  checkMenuEntries(MOCK_STORAGE.map(address =>
    JSON.stringify({
      primary: address.tel,
      secondary: FormAutofillUtils.toOneLineAddress(address["street-address"]),
    })
  ));
});
*/

</script>

<p id="display"></p>

<div id="content">

  <form id="form1">
    <p>This is a basic form.</p>
    <p><label>Name: <input id="cc-name" autocomplete="cc-name"></label></p>
    <p><label>Card Number: <input id="cc-number" autocomplete="cc-number"></label></p>
    <p><label>Expiration month: <input id="cc-exp-month" autocomplete="cc-exp-month"></label></p>
    <p><label>Expiration year: <input id="cc-exp-year" autocomplete="cc-exp-year"></label></p>
    <p><label>CSC: <input id="cc-csc" autocomplete="cc-csc"></label></p>
  </form>
</div>

<pre id="test"></pre>
</body>
</html>
